+++ lib/src/risk_system/locations/electricity_location.dart
@@ -1,7 +1,6 @@
 import 'package:date/date.dart';
 import 'package:elec/elec.dart';
 import 'package:elec/risk_system.dart';
-import 'package:elec/src/iso/iso.dart';
 import 'package:timeseries/timeseries.dart';
 import 'package:timezone/timezone.dart' as tz;
 import 'location.dart';